requirements.txt
google-api-python-client==2.111.0
pandas==2.2.3
numpy==1.26.4
streamlit==1.39.0
pyyaml==6.0.2
sqlite3
wordcloud==1.9.3
matplotlib==3.9.2